chore: update for code duplication
