+++ app/pages/analyst/cbc/[cbcId]/edit/[section].tsx
@@ -16,6 +16,7 @@ import { RJSFSchema } from '@rjsf/utils';
 import useModal from 'lib/helpers/useModal';
 import { ChangeModal } from 'components/Analyst';
 import { useUpdateCbcDataAndInsertChangeRequest } from 'schema/mutations/cbc/updateCbcDataAndInsertChangeReason';
+import { createCbcSchemaData } from 'utils/schemaUtils';
 
 const getCbcSectionQuery = graphql`
   query SectionCbcDataQuery($rowId: Int!) {
@@ -57,87 +58,7 @@ const EditCbcSection = ({
   const { cbcDataByCbcId, rowId } = cbcByRowId;
   const { jsonData, rowId: cbcDataRowId } = cbcDataByCbcId.edges[0].node;
 
-  const tombstone = {
-    projectNumber: jsonData.projectNumber,
-    originalProjectNumber: jsonData.originalProjectNumber,
-    phase: jsonData.phase,
-    intake: jsonData.intake,
-    projectStatus: jsonData.projectStatus,
-    changeRequestPending: jsonData.changeRequestPending,
-    projectTitle: jsonData.projectTitle,
-    projectDescription: jsonData.projectDescription,
-    applicantContractualName: jsonData.applicantContractualName,
-    currentOperatingName: jsonData.currentOperatingName,
-    eightThirtyMillionFunding: jsonData.eightThirtyMillionFunding,
-    federalFundingSource: jsonData.federalFundingSource,
-    federalProjectNumber: jsonData.federalProjectNumber,
-  };
-
-  const projectType = {
-    projectType: jsonData.projectType,
-    transportProjectType: jsonData.transportProjectType,
-    highwayProjectType: jsonData.highwayProjectType,
-    lastMileProjectType: jsonData.lastMileProjectType,
-    lastMileMinimumSpeed: jsonData.lastMileMinimumSpeed,
-    connectedCoastNetworkDependant: jsonData.connectedCoastNetworkDependant,
-  };
-  const locationsAndCounts = {
-    projectLocations: jsonData.projectLocations,
-    communitiesAndLocalesCount: jsonData.communitiesAndLocalesCount,
-    indigenousCommunities: jsonData.indigenousCommunities,
-    householdCount: jsonData.householdCount,
-    transportKm: jsonData.transportKm,
-    highwayKm: jsonData.highwayKm,
-    restAreas: jsonData.restAreas,
-  };
-
-  const funding = {
-    bcFundingRequest: jsonData.bcFundingRequest,
-    federalFunding: jsonData.federalFunding,
-    applicantAmount: jsonData.applicantAmount,
-    otherFunding: jsonData.otherFunding,
-    totalProjectBudget: jsonData.totalProjectBudget,
-  };
-
-  const eventsAndDates = {
-    nditConditionalApprovalLetterSent:
-      jsonData.nditConditionalApprovalLetterSent,
-    bindingAgreementSignedNditRecipient:
-      jsonData.bindingAgreementSignedNditRecipient,
-    announcedByProvince: jsonData.announcedByProvince,
-    dateApplicationReceived: jsonData.dateApplicationReceived,
-    dateConditionallyApproved: jsonData.dateConditionallyApproved,
-    dateAgreementSigned: jsonData.dateAgreementSigned,
-    proposedStartDate: jsonData.proposedStartDate,
-    proposedCompletionDate: jsonData.proposedCompletionDate,
-    reportingCompletionDate: jsonData.reportingCompletionDate,
-    dateAnnounced: jsonData.dateAnnounced,
-  };
-
-  const miscellaneous = {
-    projectMilestoneCompleted: jsonData.projectMilestoneCompleted,
-    constructionCompletedOn: jsonData.constructionCompletedOn,
-    milestoneComments: jsonData.milestoneComments,
-    primaryNewsRelease: jsonData.primaryNewsRelease,
-    secondaryNewsRelease: jsonData.secondaryNewsRelease,
-    notes: jsonData.notes,
-  };
-
-  const projectDataReviews = {
-    locked: jsonData.locked,
-    lastReviewed: jsonData.lastReviewed,
-    reviewNotes: jsonData.reviewNotes,
-  };
-
-  const dataBySection = {
-    tombstone,
-    projectType,
-    locationsAndCounts,
-    funding,
-    eventsAndDates,
-    miscellaneous,
-    projectDataReviews,
-  };
+  const dataBySection = createCbcSchemaData(jsonData);
 
   const changeModal = useModal();
